fix(thymeleaf): wrong conversion in all integer options
Linked with #289

+++ datatables-thymeleaf/src/main/java/com/github/dandelion/datatables/thymeleaf/processor/attr/TableAjaxPipeSizeAttrProcessor.java
@@ -29,6 +29,7 @@
  */
 package com.github.dandelion.datatables.thymeleaf.processor.attr;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 import org.thymeleaf.Arguments;
@@ -64,8 +65,8 @@ public class TableAjaxPipeSizeAttrProcessor extends AbstractTableAttrProcessor {
    protected void doProcessAttribute(Arguments arguments, Element element, String attributeName,
          Map<Option<?>, Object> stagingConf) {
 
-      Integer attrValue = AttributeUtils.parseAttribute(arguments, element, attributeName, Integer.class);
+      BigDecimal attrValue = AttributeUtils.parseAttribute(arguments, element, attributeName, BigDecimal.class);
 
-      stagingConf.put(DatatableOptions.AJAX_PIPESIZE, Integer.valueOf(attrValue));
+      stagingConf.put(DatatableOptions.AJAX_PIPESIZE, attrValue.intValue());
    }
 }
